chore: refactor publish pipeline around Git tags (GEA-12974)
Publishing to PyPI is now triggered via the pushing of Git tags instead
of requiring the use of the one `release` branch. This enables
flexibility in publishing prerelease versions outside of the `release`
branch and shaves one step off of the manual release process (merging
to `release`). The package to be published is derived from the Git tag
name, such as `pangea-django/v1.1.0` which would publish the
pangea-django package. This format follows what is already being done in
pangea-go.

Instructions on how to publish a new version of either pangea-django or
pangea-sdk have been added to `CONTRIBUTING.md`.

Job names have been cleaned up a little by dropping `pangea-` and
`python-` prefixes, because they did not provide much value when the
pipelines are already happening within the context of the pangea-python
repository.