pangea-sdk: run deep_verify test after dump_audit test (PAN-16742)
The deep_verify test depends on `dump.json` which is generated by the
dump_audit test, so the alphabetical order of the tests must account
for this. This was not caught in local testing because `dump.json` was
generated by previous runs of the tests.

Also replaced os usage with pathlib and removed the verify_audit test
which was simply a duplicate of the deep_verify test.